split samples in separate files
